component for list and list item all set up
