+++ app/component/entry-list-item/entry-list-item.html
@@ -1,0 +1,3 @@
+<div>
+  <h1>{{$ctrl.entry.title}}</h1>
+</div>
