+++ app/component/entry-list/entry-list.html
@@ -1,0 +1,3 @@
+<div class="entry-list" ng-init="entryListCtrl.getEntries()">
+  <ramble-entry-list-item ng-repeat="entry in entryListCtrl.list" entry="entry" />
+</div>
